Function Site with basic outline ported from 2023
