+++ pages/index.tsx
@@ -1,7 +1,6 @@
 import Footer from "@/components/footer";
 import FAQPage from "@/components/faq/faq";
 import NavBar from "@/components/nav-bar/nav-bar";
-import SponsorUsPage from "@/components/SponsorUsPage/SponsorUsPage";
 import MlhBanner from "@/components/mlh-banner/mlh-banner";
 import Head from "next/head";
 import TitleComponent from "@/components/title-components/title";
